Hacked out cluster mode
