+++ dev-packages/application-package/src/application-props.ts
@@ -116,13 +116,6 @@ export interface FrontendApplicationConfig extends ApplicationConfig {
  */
 export interface BackendApplicationConfig extends ApplicationConfig {
 
-    /**
-     * The default backend startup timeout in milliseconds. If specified here, then this value will be used as opposed to the default timeout.
-     * If the `--startup-timeout` option is defined, this configuration value has no effect. A negative values will disable the timeout.
-     * For more details, see [`startupTimeoutOption`](MasterProcess#startupTimeoutOption).
-     */
-    readonly startupTimeout?: number;
-
 }
 
 /**
