--- packages/core/src/node/cluster/worker-connection.ts
@@ -1,28 +1,0 @@
-/********************************************************************************
- * Copyright (C) 2017 TypeFox and others.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License v. 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0.
- *
- * This Source Code may also be made available under the following Secondary
- * Licenses when the conditions for such availability set forth in the Eclipse
- * Public License v. 2.0 are satisfied: GNU General Public License, version 2
- * with the GNU Classpath Exception which is available at
- * https://www.gnu.org/software/classpath/license.html.
- *
- * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
- ********************************************************************************/
-
-import { Worker } from 'cluster';
-import { WorkerMessageWriter } from './worker-writer';
-import { WorkerMessageReader } from './worker-reader';
-import { createMessageConnection, MessageConnection, Logger } from 'vscode-jsonrpc';
-
-export function createWorkerConnection(worker: Worker, logger: Logger): MessageConnection {
-    const messageReader = new WorkerMessageReader(worker);
-    const messageWriter = new WorkerMessageWriter(worker);
-    const connection = createMessageConnection(messageReader, messageWriter, logger);
-    connection.onClose(() => connection.dispose());
-    return connection;
-}
