+++ packages/core/src/node/index.ts
@@ -19,4 +19,3 @@ export * from './debug';
 export * from './file-uri';
 export * from './messaging';
 export * from './cli';
-export * from './cluster';
