+++ packages/extension-manager/src/node/test/extension-node-test-container.ts
@@ -15,8 +15,6 @@
  ********************************************************************************/
 
 import { Container } from 'inversify';
-import { stubRemoteMasterProcessFactory } from '@theia/core/lib/node';
-import { bindServerProcess } from '@theia/core/lib/node/backend-application-module';
 import { bindLogger } from '@theia/core/lib/node/logger-backend-module';
 import { bindFileSystem, bindFileSystemWatcherServer } from '@theia/filesystem/lib/node/filesystem-backend-module';
 import { ApplicationProjectArgs } from '../application-project-cli';
@@ -26,7 +24,6 @@ export const extensionNodeTestContainer = (args: ApplicationProjectArgs) => {
     const container = new Container();
     const bind = container.bind.bind(container);
     bindLogger(bind);
-    bindServerProcess(bind, stubRemoteMasterProcessFactory);
     bindFileSystem(bind);
     bindFileSystemWatcherServer(bind);
     bindNodeExtensionServer(bind, args);
