+++ packages/filesystem/src/node/filesystem-backend-module.ts
@@ -14,7 +14,6 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
 
-import * as cluster from 'cluster';
 import { ContainerModule, interfaces } from 'inversify';
 import { ConnectionHandler, JsonRpcConnectionHandler, ILogger } from '@theia/core/lib/common';
 import { FileSystemNode } from './node-filesystem';
@@ -22,6 +21,8 @@ import { FileSystem, FileSystemClient, fileSystemPath, DispatchingFileSystemClie
 import { FileSystemWatcherServer, FileSystemWatcherClient, fileSystemWatcherPath } from '../common/filesystem-watcher-protocol';
 import { FileSystemWatcherServerClient } from './filesystem-watcher-client';
 import { NsfwFileSystemWatcherServer } from './nsfw-watcher/nsfw-filesystem-watcher';
+
+const SINGLE_THREADED = process.argv.indexOf('--no-cluster') !== -1;
 
 export function bindFileSystem(bind: interfaces.Bind, props?: {
     onFileSystemActivation: (context: interfaces.Context, fs: FileSystem) => void
@@ -36,7 +37,7 @@ export function bindFileSystem(bind: interfaces.Bind, props?: {
 }
 
 export function bindFileSystemWatcherServer(bind: interfaces.Bind): void {
-    if (cluster.isMaster) {
+    if (SINGLE_THREADED) {
         bind(FileSystemWatcherServer).toDynamicValue(ctx => {
             const logger = ctx.container.get<ILogger>(ILogger);
             return new NsfwFileSystemWatcherServer({
